pyproject.toml
[tool.poetry]
name = "resumecentral"
version = "0.1.0"
description = ""
authors = ["Andrei Piscoran <aosxap@gmail.com>", "Andrei Panduru <panduru.andrei04@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.13,>=3.10"
djangorestframework = "^3.15.2"
invoke = "^2.2.0"

transformers = "^4.42.4"
requests = "^2.32.3"
langchain-text-splitters = "^0.2.2"
langchain-community = "^0.2.7"
langchain-chroma = "^0.1.2"
langchain-huggingface = "^0.0.3"
pymupdf4llm = "^0.0.9"
chromadb = "^0.5.5"
psycopg2-binary = "^2.9.9"
django-on-heroku = "^1.1.2"
langchain-core = "^0.2.23"
semantic-kernel = "^1.3.0"
ollama = "^0.3.0"
ngrok = "^1.3.0"
pyngrok = "^7.2.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
ruff = "^0.5.1"
pyright = "^1.1.371"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"